[#5] Ref: 주소검색화면에서 Model(로직,데이터) 분리
